+++ packages/react/src/index.ts
@@ -8,4 +8,7 @@ export * from './hooks/useRealtime';
 export * from './hooks/useFiles';
 export * from './hooks/useInvitation';
 export * from './hooks/useClientAction';
+export * from './hooks/useCodelistValues';
+export * from './hooks/useCodelists';
+export * from './hooks/useCodelistOverride';
 export * from './types';